fix(nested_dissection): avoid unsigned overflow in balance computation
node_count*min_balance is evaluated in 32-bit unsigned arithmetic and
wraps once node_count exceeds 2^32/min_balance (about 107M nodes for
the 40%-balance cutter), silently shrinking side_size to a tiny or even
zero value. The result is an arbitrarily unbalanced cut at the top
recursion levels of planet-scale graphs: worse nested-dissection
orders and deeper recursion, with no assertion firing.

Evaluate the product in 64-bit, as the cut-quality comparisons in
the three-cut inertial_flow overload already do.

+++ src/nested_dissection.cpp
@@ -503,7 +503,7 @@ CutSide inertial_flow(
 
 	const unsigned node_count = g.node_count();
 
-	unsigned side_size = (node_count*min_balance)/100;
+	unsigned side_size = static_cast<unsigned long long>(node_count)*min_balance/100;
 	if(side_size == 0)
 		side_size = 1;
